+++ CHANGES.rst
@@ -2,6 +2,12 @@
 Changes
 =======
 
+development version
+-------------------
+
+* Add a ``config`` subcommand, which changes the configuration file
+  from the command-line.
+
 v0.6.0 (2016-12-07)
 -------------------